Extract all step loading to StepLoader
