+++ lib/turnip/loader.rb
@@ -3,7 +3,7 @@ module Turnip
     def load(*a, &b)
       if a.first.end_with?('.feature')
         require 'spec_helper'
-        Turnip::Config.load_steps
+        Turnip::StepLoader.load_steps
         Turnip.run(Turnip::FeatureFile.new(a.first))
       else
         super
